+++ Matrix_solve/utils.h
@@ -10,7 +10,7 @@ extern "C"
 
 
 int row_of_zero(int m, int n, double* *matrix, int check_row);
-void set_matrix(double* *matrix, int val_row, int val_col, int val);
+void set_matrix(double* *matrix, int val_row, int val_col, double val);
 void delete_matrix(int m, double* *matrix);
 
 #ifdef __cplusplus
@@ -18,3 +18,4 @@ void delete_matrix(int m, double* *matrix);
 #endif
 
 #endif /* utils_h */
+
